formula to implement
Length of each side is 1, and each land has 4 sides, but if two lands are adjacent, length of two edges should be removed. So, the perimeter equals to number of land times 4 minus number of internal edges times 2. Traverse all grids, count the number of lands and edges. For each cell, if it is 1, which means it is a land, add 1 to number of lands. For each land, if there is a land at its left side, add 1 to number of edges, and similarly, if a land exists above, add 1 to number of edges.
Alternative approach can be using DFS